requirements.txt
asgiref==3.6.0
Django==4.2
django-cleanup==6.0.0
django-encrypted-field==1.0.3
django-robots==5.0
django-searchable-encrypted-fields==0.2.0
mysqlclient==2.1.1
pycryptodome==3.17
pycryptodomex==3.17
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3